Add new pubkey for Alert Messages
The original pubkey for Alert Messages was inserted by original developer(s)
which have long since abandonded the project. No one has the private key
for that pubkey, so we cannot send any alert messages.
The new key has not been tested yet, but the exisiting one doesn't work
anyway. sxcer created the key pair and it will be distributed to several
users at sexcoinforum.org.

As a sidenote, the original pubkey is identical to the one distributed
with the last known RoyalCoin sources:

https://sourceforge.net/projects/royalcoin/files/royalcoin-source.zip/download